Backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.7.2
async-timeout==4.0.3
autopep8==2.0.4
bitarray==2.9.2
ckzg==2.0.0
cloudinary==1.36.0
cytoolz==0.12.3
distlib==0.3.8
dj-database-url==2.1.0
dj-rest-auth==5.0.2
Django==5.0.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
filelock==3.14.0
frozenlist==1.4.1
hexbytes==1.2.1
inflection==0.5.1
lru-dict==1.3.0
multidict==6.0.5
packaging==23.2
parsimonious==0.10.0
paystackapi==2.1.1
pillow==10.2.0
platformdirs==4.2.2
psycopg==3.1.17
pycodestyle==2.11.1
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
python-decouple==3.8
pyunormalize==15.1.0
regex==2024.7.24
rlp==0.6.0
safe-pysha3==1.0.4
semantic-version==2.10.0
sqlparse==0.4.4
tomli==2.0.1
toolz==0.12.1
types-requests==2.32.0.20240712
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.2
web3==3.16.5
websockets==13.0.1
yarl==1.9.7
